+++ Tests/SwiftyModelTests/RelationsTests/ToOneTests.swift
@@ -1,0 +1,8 @@
+//
+//  File.swift
+//  
+//
+//  Created by Sergey Kazakov on 13/07/2024.
+//
+
+import Foundation
